Change function to filter highest tech engies, not assist them
As it turns out, the previous commit message was not true as the
function does not work correctly without a callback. The selection part
works, but the assisting of a single highest tech engineer doesn't. It
seems like UI functions don't have access to the sim/c++ functions that
are needed to give assist ("guard") commands and that's what the
simcallback was for, being basically a wrapper for this function to be
called by mods. For some reason I'm unable to hook into this file,
import it or change the table that defines the callback function in any
way from the mod directory. This might be for security reasons, or I'm
just doing it wrong.
For now I've changed the function to just select all highest tech engies
in the selection, which is fine for most use cases, but what I'd really
want is that and each lower tech engie to then assist their closest
(or any) highest tech engie that's nearby. Giving such an assist/guard
command would be easy if I could get the callback to work or could call
the sim functions directly.
Should ask the more experienced modders about that.